added request handler

+++ back-end/src/api/controllers/customer_controller.py
@@ -16,7 +16,6 @@ def get_customers():
 
     if req_handler_response['success']:
         core_app_context = req_handler_response['core_app_context']
-        # the core app context instance can be passed as argument in the following method
         api_response = customer_service.get_customers()
         return api_response
 
@@ -27,36 +26,59 @@ def get_customer(customer_id):
     """
         This is the endpoint returning a single customer with the given id
     """
+    req_handler_response = request_handler.validate_token(request)
 
-    api_response = customer_service.get_customer(customer_id)
-    return api_response
+    if req_handler_response['success']:
+        core_app_context = req_handler_response['core_app_context']
+        api_response = customer_service.get_customer(customer_id)
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/', methods=['POST'])
 def insert_customer():
     """
         This is the endpoint for creating a new customer
     """
-    data = request.get_json()
-    api_response = customer_service.insert_customer(data)
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        core_app_context = req_handler_response['core_app_context']
+        data = request.get_json()
+        api_response = customer_service.insert_customer(data)
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/', methods=['DELETE'])
 def delete_customer():
     """
         This is endpoint for deleting a customer
     """
-    data = request.get_json()
-    api_response = customer_service.delete_customer(data)
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        core_app_context = req_handler_response['core_app_context']
+        data = request.get_json()
+        api_response = customer_service.delete_customer(data)
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/', methods=['PUT'])
 def update_customer():
     """
         This is endpoint for updating a customer
     """
-    data = request.get_json()
-    api_response = customer_service.update_customer(data)
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        core_app_context = req_handler_response['core_app_context']
+        data = request.get_json()
+        api_response = customer_service.update_customer(data)
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/<customer_id>/make-active', methods=['PUT'])
 def make_active(customer_id):
@@ -64,8 +86,14 @@ def make_active(customer_id):
         This is endpoint for updating the is_active attribute
         of the given customer to true
     """
-    api_response = customer_service.make_active(customer_id)
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        core_app_context = req_handler_response['core_app_context']
+        api_response = customer_service.make_active(customer_id)
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/<customer_id>/make-passive', methods=['PUT'])
 def make_passive(customer_id):
@@ -73,5 +101,11 @@ def make_passive(customer_id):
         This is endpoint for updating the is_active attribute
         of the given customer to false
     """
-    api_response = customer_service.make_passive(customer_id)
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        core_app_context = req_handler_response['core_app_context']
+        api_response = customer_service.make_passive(customer_id)
+        return api_response
+
+    return req_handler_response
